Update readme project title, move dumb components
