+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
-import App from './components/App';
+import App from './container/App';
 import rootReducer from './reducers/index';
 import './index.css';
