+++ src/reducers/mainFilter.js
@@ -3,7 +3,6 @@ import { FILTER_UPDATE } from '../helpers/help';
 const mainFilter = (state = 'market_cap_desc', action) => {
   switch (action.type) {
     case FILTER_UPDATE:
-      // console.log(`@mainFilter reducer, filter is : ${action.filter}`);
       return action.filter;
 
     default:
